Remove hack, add more output
Remove the hack that we had in place to not run modules with charges
first cycle.  Now everything runs inside the loop, so much cleaner.
Cache the results of each run, so we can output it and return it back in
case we want to do something like graph it out.